Compartmentalize all. the. functions. \o/
