+++ modules/areas.js
@@ -1,0 +1,46 @@
+const AREAS = {
+  'tas': [
+    'hobart', 'hobart airport', 'bushy park', 'campania', 'cape bruny', 'dennes point',
+    'dover', 'dunalley', 'grove', 'kunanyi', 'maria island', 'orford', 'tasman island'
+  ],
+  'act': [
+    'canberra', 'batemans bay', 'braidwood', 'cabramurra', 'cooma', 'goulburn airport', 'montague island',
+    'moruya airport', 'mount ginini', 'narooma', 'nowra', 'perisher valley', 'thredbo top station',
+    'tuggeranong', 'ulladulla'
+  ],
+  'nsw': [
+    'sydney', 'sydney airport', 'sydney harbour', 'olympic park', 'badgerys creek', 'bankstown', 'bellambi',
+    'camden', 'cambletown', 'canterbury', 'fort denison', 'gosford', 'holsworthy', 'horsley park',
+    'katoomba', 'kurnell', 'lake macquarie cooranbong', 'little bay', 'lucas heights', 'mangrove mountain',
+    'mount boyce', 'newcastle', 'norah head', 'north head', 'parramatta', 'penrith', 'prospect dam',
+    'richmond', 'springwood', 'terrey hills', 'wattamolla', 'williamtown'
+  ],
+  'vic': [
+    'melbourne', 'melbourne airport', 'avalon', 'bundoora', 'cerberus', 'coldstream', 'cranbourne',
+    'essendon airport', 'fawkner beacon', 'ferny creek', 'frankston', 'geelong racecourse', 'laverton',
+    'moorabbin airport', 'phillip island', 'point wilson', 'rhyll', 'scoresby', 'sheoaks',
+    'south channel island', 'st kilda harbour rmys', 'viewbank'
+  ],
+  'sa': [
+    'adelaide', 'adelaide airport', 'edinburgh', 'hindmarsh island', 'kuitpo', 'mount barker',
+    'mount crawford', 'mount lofty', 'murray bridge', 'noarlunga', 'nuriootpa', 'outer harbor black pole',
+    'pallamana', 'parafield', 'parawa west', 'roseworthy', 'sellicks hill', 'strathalbyn', 'turretfield', 'victor harbor'
+  ],
+  'qld': [
+    'brisbane', 'brisbane airport', 'amberley', 'archerfield', 'banana bank', 'beerburrum', 'beaudesert aws',
+    'cape moreton', 'coolangatta', 'double island point', 'gratton', 'gold coast seaway', 'gympie',
+    'inner beacon', 'jimna', 'kingaroy', 'logan city', 'maroochydore', 'nambour', 'oakley', 'point lookout',
+    'rainbow beach', 'redcliffe', 'redland alexandra hills', 'spitfire channel', 'tewantin', 'toowoomba',
+    'warwick', 'wellcamp airport'
+  ],
+  'wa': [
+    'perth', 'perth airport', 'bickley', 'garden island', 'gooseberry hill', 'gosnells', 'hillarys point boat harbour',
+    'jandakot', 'mandurah', 'medina', 'melville water', 'ocean reef', 'pearce', 'rottnest island', 'swanbourne'
+  ],
+  'nt': [
+    'darwin', 'darwin harbour', 'batchelor', 'black point', 'channel point', 'charles point', 'dum in mirrie',
+    'middle point', 'noonamah', 'pirlangimpi', 'point fawcett', 'point stuart'
+  ]
+};
+
+module.exports = AREAS;
