+++ test/colorTemp.test.js
@@ -1,0 +1,15 @@
+// const colorTemp = require('./../modules/colorWind');
+//
+// describe('with a string', () => {
+//   describe('of 51.1', () => {
+//     test('it returns a color formatted string', () => {
+//       expect(colorTemp('51.1')).toEqual('51.1');
+//     });
+//   });
+//
+//   describe('of 24.1', () => {
+//     test('it returns a color formatted string', () => {
+//       expect(colorTemp('24.1')).toBe('24.1');
+//     });
+//   });
+// });
